Start queries refactor
