+++ pgqueuer/types.py
@@ -4,18 +4,9 @@ from typing import Literal, NewType
 
 ###### Events ######
 PGChannel = NewType("PGChannel", str)
-OPERATIONS = Literal[
-    "insert",
-    "update",
-    "delete",
-    "truncate",
-]
-
-EVENT_TYPES = Literal[
-    "table_changed_event",
-    "requests_per_second_event",
-    "cancellation_event",
-]
+OPERATIONS = Literal["insert", "update", "delete", "truncate"]
+STATUS = Literal["queued", "picked"]
+EVENT_TYPES = Literal["table_changed_event", "requests_per_second_event", "cancellation_event"]
 
 
 ###### Jobs ######
